server/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-ollama
httpx
python-dotenv
